Move HopsSelector into header, TODO: Header tests
